+++ src/main/java/com/pocopi/api/models/test/TestQuestionModel.java
@@ -32,7 +32,7 @@ public class TestQuestionModel {
     private TestPhaseModel phase;
 
     @Column(name = "`order`", nullable = false, columnDefinition = "int1 unsigned")
-    private byte order;
+    private short order;
 
     @Builder.Default
     @Size(min = TEXT_MIN_LEN, max = TEXT_MAX_LEN)
